feat(#400): display grid and axis options only if needed
